more work on details screen

+++ lib/model/movie_details_response.dart
@@ -22,7 +22,7 @@ class MovieDetailsResponse {
   String tagline;
   String title;
   bool video;
-  int voteAverage;
+  double voteAverage;
   int voteCount;
 
   MovieDetailsResponse(
